move classes back to cpp file and add default identity op
