+++ src/targets/gpu/include/migraphx/gpu/prepare_reduce.hpp
@@ -27,10 +27,7 @@
 
 #include <migraphx/config.hpp>
 #include <migraphx/gpu/export.h>
-#include <migraphx/operation.hpp>
-#include <migraphx/shape.hpp>
 #include <string>
-#include <algorithm>
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
@@ -38,66 +35,6 @@ inline namespace MIGRAPHX_INLINE_NS {
 struct module;
 
 namespace gpu {
-
-struct parallel_reduce
-{
-    operation op;
-
-    template <class Self, class F>
-    static auto reflect(Self& self, F f)
-    {
-        return pack(f(self.op, "op"));
-    }
-
-    std::string name() const { return "gpu::parallel_reduce"; }
-
-    shape compute_shape(const std::vector<shape>& inputs) const
-    {
-        std::vector<shape> result;
-        std::transform(inputs.begin(), inputs.end(), std::back_inserter(result), [&](auto input) {
-            return op.compute_shape({input});
-        });
-        return shape{result};
-    }
-};
-
-struct arg_reduce
-{
-    operation op;
-
-    template <class Self, class F>
-    static auto reflect(Self& self, F f)
-    {
-        return pack(f(self.op, "op"));
-    }
-
-    std::string name() const { return "gpu::arg_reduce"; }
-
-    shape compute_shape(const std::vector<shape>& inputs) const
-    {
-        auto index_shape = op.compute_shape({inputs.front()});
-        auto value_shape = index_shape.with_type(inputs.front().type());
-        return shape{{value_shape, index_shape}};
-    }
-};
-
-struct make_indices
-{
-    std::size_t size = 0;
-
-    template <class Self, class F>
-    static auto reflect(Self& self, F f)
-    {
-        return pack(f(self.size, "size"));
-    }
-
-    std::string name() const { return "gpu::make_indices"; }
-
-    shape compute_shape(const std::vector<shape>&) const
-    {
-        return shape{shape::uint32_type, {size}};
-    }
-};
 
 struct MIGRAPHX_GPU_EXPORT prepare_reduce
 {
